Issue #6622: no final methods: Consistent (and shorter) method names, and use long consistently
Since TreeMaker.Modifiers take a long we should just use long everywhere

+++ compiler-java/src/com/redhat/ceylon/compiler/java/codegen/ClassDefinitionBuilder.java
@@ -573,11 +573,11 @@ public class ClassDefinitionBuilder
         return companionBuilder;
     }
 
-    public ClassDefinitionBuilder field(int modifiers, String attrName, JCExpression type, JCExpression initialValue, boolean isLocal) {
+    public ClassDefinitionBuilder field(long modifiers, String attrName, JCExpression type, JCExpression initialValue, boolean isLocal) {
         return field(modifiers, attrName, type, initialValue, isLocal, List.<JCTree.JCAnnotation>nil());
     }
     
-    public ClassDefinitionBuilder field(int modifiers, String attrName, JCExpression type, JCExpression initialValue, boolean isLocal, 
+    public ClassDefinitionBuilder field(long modifiers, String attrName, JCExpression type, JCExpression initialValue, boolean isLocal, 
             List<JCTree.JCAnnotation> annotations) {
         if (!isLocal) {
             // A shared or captured attribute gets turned into a class member
